Move common request type from providers package to its own package.
