+++ formatters/table.go
@@ -5,14 +5,14 @@ import (
 	"time"
 	"unicode/utf8"
 
-	"github.com/cego/git-request-list/providers"
+	"github.com/cego/git-request-list/request"
 )
 
 // Table represents an ASCII table containing pull- and merge-requests.
 type Table struct{}
 
 // String returns the ASCII table t containing the given requests.
-func (t *Table) String(requests ...providers.Request) string {
+func (t *Table) String(requests ...request.Request) string {
 	rows := [][]string{{"Repository", "Name", "URL", "Created", "Updated"}}
 	for _, r := range requests {
 		rows = append(rows, []string{r.Repository, r.Name, r.URL, r.Created.Format(time.UnixDate), r.Updated.Format(time.UnixDate)})
